Change where DataService triggers store their data
DataService triggers previously stored data for properties they were
managing in a separate prototype chain. One consequence of this was that
the "this" value in getters and setters would be different depending on
whether a property was managed by a trigger or not.

This commit fixes that by making triggers use the same prototype chain
used in non-triggering getters and setters. For a property named "X",
triggers will store their values for that property in "_X".

This commit also extracts the trigger functionality of DataServices
to a separate private DataTrigger class.