Reorder RawDataService sections & update comments
The sections were reordered to match the DataService section order.

Comments were clarified and updated to the latest expected behavior.

The unused saveDataChanges() method was removed, to be replaced at some
point in the future by a saveData() method that would also replace
saveDataObject(), at the same time as deleteDataObject() would be
replaced by deleteData() and createData(), readData(), and updateData()
would be added.

Also, the logic for determining which service is the main service was
tweaked slightly, from the root of the last created service to the root
of the first created service. This shouldn't change anything for most
applications where all services have the same root, but this is more
deterministic and easier to understand. Some tests were failing with
the previous logic and now pass.