First attempt data mapping.
Create subclass of data mapping called blueprint data mapping.  If raw
data service has a blue print for the object and mapping is not defined
it will create a blueprint data mapping.

Dependent on a change property blueprint.  The property blueprint needs
a new property called "synonym".  This is an optional property that
provides a way to map properties of raw data to object e.g.

object.firstName = data.GIVEN_NAME