Create work around for Safari 10.1 bug
If the user's browser is Safari and it is version 10.1 the application
needs to copy the contents of the database into a temp database and then
reimport the data.  This is a one-time operation.

Bugs outlined here:

The indexedDB bug described here:

https://bugs.webkit.org/show_bug.cgi?id=170358
and here: https://bugs.webkit.org/show_bug.cgi?id=170000